requirements.txt
# M5Stack IoT Gateway Dependencies

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# MQTT Client
paho-mqtt==1.6.1

# HTTP Requests
requests==2.31.0

# Type Hints (Python < 3.9)
typing-extensions==4.7.1

# Development Dependencies (optional)
# pytest==7.4.2
# pytest-flask==1.2.0
# black==23.7.0
# flake8==6.0.0
